write createBlockIndex function and replace code in era1
